d3d12; Live buffer related code and factorise resource_desc for buffers
